Add continuous integration jobs (#1)
When we moved to this repository from the jlm-eval-suite repository, we
dropped the CI jobs that were present in jlm-eval-suite. This PR fixes
this regression by reintroducing the jobs in this repository. However,
the jobs are introduced with a convenient way to disable them by simply
appending `[DisableCI]` to the PR title. This enables us to break the
cycle between the jlm repository and this repository when we, for
example, have to update the LLVM version.