Make jlm/Makefile.sub build system agnostic. (#6)
Until now `jlm/Makefile.sub` assumed that Unix Makefiles was configured
as the default generator for Cmake. On systems where this was not the
case, the build step failed. This PR switches over to using `cmake
--build` for the build step instead of the previously hardcoded `make`.
All targets are now also defined in a uniform way without the manual
`mkdir` and `cd` calls.

Behavior on systems with `CMAKE_GENERATOR='Unix Makefiles'` should be
unchanged. On other systems, the script will no longer fail but use the
configured generator.